fsck: rework the per-packfile logic in do_pack
Perhaps easier to follow, and now always runs both git and par2
verifications when possible since git verification is checking the
internal correctnes of the packfile, while par2 is just making sure
that whatever it was given during generate hasn't changed.  If git
verify and par2 verify disagree, just exit with an error for now.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>