Wait for btrfs/zfs sparse info sync when there's no SEEK_DATA/HOLE
NetBSD (for example) doesn't currently have support for lseek
SEEK_DATA/SEEK_HOLE, so detect when that's missing and fall back to
du, and since we've learned that the du output may take up to 5s to
settle on some filesystems like btrfs and zfs, just sleep
unconditionally before checking for now.

To support this, move the filesystem type reporting from dev/lib.sh to
dev/path-fs, and augment dev/sparse-size.

Thanks to Greg Troxel for reporting the issue, and proposing/testing
the sleep approach.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>