walk_object: make it possible to always notice missing blobs
Allow providing an oid_exists function to walk_object.  When provided,
use it to check the existence of blobs when include_data is false.
Previously, those blobs would just be returned as if they existed,
based on the information from the parent tree.  This will support
forthcoming changes to detect incomplete trees.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>