walk_object: walk depth first (don't let get save incomplete trees)
Yield children before parents so that bup-get(1) never stores
incomplete treeish objects in the repo.  Otherwise it might leave them
in the repository if it were interrupted after it stored a tree and
before it stored all of the descendants.

Previously, given

  ... ---- tree-1 +---- blob-1
                  +---- blob-2

walk_object() would yield the objects in this order:

  - tree-1
  - blob-1
  - blob-2

If none of these objects existed in the target repository, and if bup
get were interrupted just after it had stored blob-1, then the
repository would be left with an incomplete tree-1 (i.e. with blob-2
missing).  Running bup get again won't fix the problem because it will
stop as soon as it sees that tree-1 already exists in the destination.
This commit changes the order to:

  - blob-1
  - blob-2
  - tree-1

so that the respostory never includes an incomplete tree.

Thanks to Johannes Berg for providing an earlier variant of this fix
and enhancements to the commit message.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>
(cherry picked from commit d530483989ab2dd8c33d58774ccab9980e84d1a6)