Only use config options in REPO/config
Even if bup stores data in the same way as git, bup still isn't git.
Therefore, it doesn't make sense for bup to use the /etc/gitconfig
or ~/.gitconfig, they should have different settings.

There are also bup specific settings such as bup.split.trees or the
upcoming bup.split.method, that could result in confusing results
if (accidentally or not) set in a git config file that's outside the
repository. For example, mounting a repo on an external drive on
machines with different configurations could result in it not using
deduplication (efficiently).

Therefore, always use "git config --file" to avoid it using values
from locations other than the bup repository configuration file.

Note that this technically is a change in behaviour for people who
have settings that are relevant to bup in a file that git would have
read before this change, but the only setting is pack.packSizeLimit
which doesn't seem too likely to have been used globally.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: add release note]
[rlb@defaultvalue.org: simplify git_config_get and repo_config_file]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>