fsck: remove --no-rev-index (breaks supported git versions)
Remove the --no-rev-index argument that older versions don't support,
and thus fail, e.g. on Debian oldstable/bullseye.

Also, the setting is pretty harmless, with it you just get a .rev file
for each pack file.

We could (automatically or recommend) setting in the config file
pack.writeReverseIndex=false, so that it doesn't happen, which would
be ignored by older git.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: adjust commit message]
Tested-by: Rob Browning <rlb@defaultvalue.org>