Raise minimum make version to 4.2 for SHELLSTATUS
Since the GNUmakefile uses .SHELLSTATUS, which is only in make since
4.1.90, bump the minimum requirement to 4.2 (it's easier to not have
to parse three levels, and all relevant systems have it anyway.)

Also make MacOS use gmake in CI, the make there is GNU, but only
version 3.81.

For FreeBSD also fix /bin/bash to /usr/bin/env bash.

Also print out the version string if it fails.

The expr version comparison in validate-make is also incorrect, since
string comparisons (which expr does if the values aren't integers)
result in e.g. 3.9 > 3.81, but we'll fix that separately.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: omit expr comparison bug fix]
[rlb@defaultvalue.org: adjust commit message]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>