test_bloom: increase false positives threshold
This is a random test, and it occasionally fails. According to git
history, it has been this way since the introduction of bloom, across
many changes. Since it's random, there's really no way to ensure it
cannot fail; however, bump the number of allowed false positives up on
the assumption that real breakage in the bloom code would be
catastrophic, and not manifest in a small deviation from the false
positive rate.

This makes this test significantly more likely to pass; it wasn't
likely to fail before, but did sometimes.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>