missing-objects: switch to dot graphs; render to html
Replace the mermaid graphs with graphviz dot (presumably more easily
portable), render those as svg, and then include them in an html
version of the document.  Adjust the document to reflect the
differences, add a title and trivial intro, etc.

Move "how does it happen" a bit further down and cross-reference.
Replace a couple of graphs with a bit more prose.

If pandoc and dot are available, you should be able to build

For now, check in the generated html so that it will always be
available, say via "firefox issue/missing-objects.html".

Signed-off-by: Rob Browning <rlb@defaultvalue.org>