gc: add --ignore-missing
Add an --ignore-missing option that allows a gc to proceed even if it
encounters missing objects.  For now just log the paths to those
objects informally, and exit with nonzero status if any are
discovered.

This will allow a gc on a damaged repository while we're still working
on a repair method -- both to reclaim space (as usual) and to purge
the repository of hazardous orphaned incomplete trees or
commits (cf. issue/missing-objects.md).

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>