Add bup validate-ref-links
Scan all the live data and log the paths to any missing objects.  This
is essentially what the first phase of a gc does (and it shares the
same code).  This is intended to provide a read-only integrity check
and a way to see what may have been lost if any objects are missing.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>