Allow specification of refs for validate-ref-links
Make it possible to scan only a specified set of references.  This is
a more efficient way to check the integrity of saves than either
joining them to /dev/null or restoring them, particularly if the
references share much data via deduplication.

Note that it does not actually read any of the (leaf/blob) data, so it
is a bit less thorough than a join or restore.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>