Simplify PackMidx by requiring construction via open_midx
Rework PackMidx __init__() to only handle valid
midxes (i.e. MIDX_VERSION) and for now to be "internal only".  Move
the version checking and other validation to a new open_midx()
function that returns None for invalid mixes or throws MissingIdxs if
any of the consituent indexes are missing *and* ignore_missing is
false.  This eliminates the need for both the force_keep and
_init_failed() related states.

This should also ensure that remote operations (via client) like save
involving a repository that has changed (e.g. has been gc'ed) don't
risk deciding that objects exist in the remote repository when they're
actually gone.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>