Move make version check to GNUmakefile; drop dev/validate-make
Move the check to the makefile since that's the only place we need it,
and since it's feasible to handle it there (and simpler), particularly
in the current situation where we only need to make sure the major
version is at least 4, and the minor version isn't 0 or 1.

This avoids the expr comparison problem mentioned in

  Raise minimum make version to 4.2 for SHELLSTATUS
  8d0d32dc69251f59a636d7748c62ff8c43fd6f6c

i.e. that string comparisons (which expr does if the values aren't
integers) are incorrect, e.g. 3.9 > 3.81).

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>