midx: fix path when checking idx in midx
If a specific midx file in the current working dir is checked, the
path is calculated in the wrong way and the code attempts to look up
with idx file in / rather than the current working dir. Fix that.

Thanks to Greg Troxel for reporting the issue.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: add release note]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>
(cherry picked from commit 164bd3ae9bab3164aab9719312266ef7f06bdccc)