main: don't close filter pipes from another thread
This, of course, relies on undefined behavior, and it can cause
commands to hang waiting on the filter thread to finish when closing
the pipe doesn't cause the filter select() to return.

Instead, add a control pipe, and write "q" (quit) to that to notify
the filter thread that it should exit.  Adjust the cleanup behavior
and rework it to use an ExitStack.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>