progress: clear to end-of-line when appropriate
When the previous progress line ends in a carriage return, issue an
ANSI "Erase in Line" to clear the line before printing the new
message.

  https://en.wikipedia.org/wiki/ANSI_escape_code#Control_Sequence_Introducer_commands

Also call progress('') in main after every python subcommand to make
sure we clear the line for the next command. Previously, if your shell
prompt (PS1) was short enough, it would intermingle with bup's last
progress line for commands like validate-object-links.

Thanks to Johannes Berg for the ANSI approach.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>