bup init: "refresh" repositories more carefully
When run on an existing repository, continue make sure the
pack.indexVersion is 2, but rather than always resetting it, only
change it if it isn't already 2.

Don't touch core.logAllRefUpdates at all, so that if it's been
intentionally removed or set to false, it stays that way.

Detect whether or not the repository "exists" similarly to the way
that git appears to, i.e. if the repository has a HEAD.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>