vfs: don't replace (dir) metadata size None with 0
If we don't know the size, don't pretend we do know it's 0.  Leave it
None and requre clients to decide how to handle it.  This matters for
'bup rewrite' which would otherwise always fill in "0" for directories
that should have "None" for their size (i.e. not stored.), changing
their hashes.

Add a new "public" argument to augment_item_meta() and
ensure_item_has_metadata() that when true, produces metadata intended
for public consumption (i.e. via ftp/ls/web). For now, that just means
that directory sizes will be 0 instead of None.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: add/use public argument]
[rlb@defaultvalue.org: adjust commit message]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>