bloom/midx: don't always check_repo_or_die()
If a dir to operate in is specified then don't require a repo. With no
longer always having a default repo this is required for working with
remote repos in case there's no local default repo.

Also remove the repo_rel() call in ruin_bloom() since that's just used
for printing but won't work without a call to check_repo_or_die().

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>