Separate compilation and linking of C code
Build .o files first, with different versions from bup.c for the
relevant executables, and then link the executables and the
library. This way, the .d files can be used properly to rebuild the
necessary things when a header file changes.

Along the way also rename "shared" to "common" as it wasn't
immediately clear to me what "shared" meant, is that shared between
the things we build (yes) or for the "shared object".

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: add bup_config_ldflags_so]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>