Test HashSplitter handling of short reads
The HashSplitter can get short reads from both file descriptor and
python objects. Add two brief tests to ensure such data is merged into
a single blob and not treated as multiple. This was never broken but
for a while I was afraid it was, so testing it seemed prudent.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>