Replace tree_decode with tree_iter and tree_entries
Rename tree_decode() to tree_iter() to make its purpose clearer, and
switch to tree_entries() in cases where a list is preferable,
e.g. where we know the trees won't be large enough to matter, and/or
we want to avoid iterator-related mistakes like this:

  396735cb814d33f994a4c7134c65c06fdae5c0d6
  vfs: tree_items: use a list for tree items (split trees)

For data created with bup.split.trees set to true, we shouldn't have
any large trees, but that won't be the case for all the existing data,
so stick with a generator for now where that might be relevant and
advantageous, i.e. vfs.join(), walk_objects(), and
validate-object-links.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>