Don't provoke duplicate index suggestions during 'bup on get ...'
When walk_object switched to a depth first traversal, get's
get_random_item() should have changed to re-check whether the writer
already has a tree whenever walk_object finally yields them, since the
client might have learned about the tree's existence during the depth
first traversal.

That would be because the client synced an index to the index-cache
that happens to include the tree in response to a server index
suggestion that was provoked by an attempt to write some other oid
that the server already has.

cf.

  walk_object: walk depth first (don't let get save incomplete trees)
  d530483989ab2dd8c33d58774ccab9980e84d1a6

Thanks to Johannes Berg for help diagnosing the problem and suggesting
the fix.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>