Move bup.split.trees into hashsplitter config
The hashsplitter config is now expected to include anything that could
cause a save to "meaningfully" differ from another created with a
different config (e.g. is tree splitting enabled, what's the split
"granularity", ...).

This will support a forthcoming "get" option to --rewrite saves when
transferring them; we're going to rely on comparing configs to
categorize each request. e.g. if the configs differ, then the transfer
should probably be a --rewrite.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>