get: support --missing fail/ignore; remap --ignore-missing
Move the missing configuration from opt to the specs since, even
though it's currently (and maybe always will be) global, it's more
accurately a property of the requested transfer, and doing so also
avoids needing extra function arguments to carry that information when
we already have the spec.

Store the missing config in a new MissingConfig dataclass that
supports 'fail' and 'ignore' modes. Change the Spec to a dataclass so
that we can more easily allow missing to default to None until we're
ready to fill it in.

Move validation of the missing config to the method specific
resolvers since they're the domain experts.

All of this prepares for support of repository repairs via "--missing
replace...".

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>