test-install: don't require pandoc
The test-install test installs bup and then runs a few checks on the
installed bup, including a call to `bup help save` which should show the
bup-save manpage, and the test greps the output for text from said page.

In environments lacking pandoc, the manpages are neither generated nor
installed, and as such, the test fails to find the expected output.

It is a common desire to install a minimal package which excludes things
like docs, and in such scenarios, the tests may still be expected to
pass.

This commit updates the test to check if pandoc is present, and if not,
skip the normal test.

Signed-off-by: Nathaniel <I@nathaniel.land>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
[rlb@defaultvalue.org: replace which with posix type -p]
[rlb@defaultvalue.org: wvskip when pandoc isn't installed]
Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>