test: avoid potential SIGPIPE failures via head -1
"something | head -1" can fail if head exits "early", causing
something to receive a SIGPIPE (and coreutils head may).

Add a WVPIPE that acts exactly like WVPASS, except that it ignores an
exit status of 128 + SIGPIPE. This is bash specific because bash
promises that the exit status of a command that's terminated by a
signal will be 128 + signal value.

Use WVPIPE to avoid the failures, and add a btl-ent-oid to extract the
oid from a git ls-tree line to be more descriptive and replace
multiple cut invocations.

Thanks to Nathaniel Hourt for reporting the problem and Nathaniel
Hourt and Johannes Berg for tracking down the cause, and Joannes Berg
for proposing the WVPIPE approach.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>