rewrite: avoid redundant work (e.g. vfs resolves)
When walking the source tree, yield paths that have the correct/full
metadata for all the items, including the directory itself if the path
is a directory. This allows us to drop the calls to resolve() the
metadata.

When we encounter a directory that's already been converted, just
provide the oid so we can call append_to_current() directly to add it
to the parent.

Thanks to Johannes Berg for investigating and diagnosing the problem
and then devising a similar solution that lead directly to this one.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>