wvtest-bash.sh: handle signal.SIGPIPE printing across py versions
signal.SIGPIPE comes from an IntEnum and looks like python changed the
print (str) behavior in 3.11:

  https://docs.python.org/3/library/enum.html#enum.IntEnum

Older python versions print Signals.SIGPIPE instead of the value.

Explicitly request the value to avoid the problem, and also add a
missing exit status check so that we fail immediately if the
arithmetic expression is invalid.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>