main: fix filter_output with subprocess subcommands
Previously any external/subprocess commands like import-rdiff-backup
would crash when isatty because filter_output was invoked without the
control argument in run_subproc_command. Make the control argument
optional.

Adjust the import-rsnapshot tests in test-misc to expose the original
problem by running the subcommand with a tty via script when it's
available (which appears likely, since it's a native bsd command, and
in the essential bsdutils package for debian derivatives).

The issue was introduced by

  39aced66ad1a6438c4ea9463621a2fa18120eb74
  main: don't close filter pipes from another thread

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>