get: make relevant arguments contextual
Make it so that each requested transfer respects the relevant
arguments that precede it on the command line, so that, for example,
this is no longer an error:

  bup get --ignore-missing --unnamed REF --no-ignore-missing --pick REF

This also applies for --exclude*, --rewrite, --missing, etc., and
whenever the excludes change, clear the rewrite db because existing
trees may have become invalid.

We'll update the docs, etc. after some additional upcoming changes to
the arguments.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>