Improve gc/validate-ref-links scan progress information
The previous information could be completely misleading when a subset
of refs was given since the progress was always/only based on the
total number of objects in the repository.

Exapnd the progress information to make it clear that the object count
is all objects, and include information about far we've gotten through
the refs that are being scanned, e.g.

  scanned 9 of 742 refs (28.32% of all objects)

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>