vfs: return restricted modes for empty (lost) metadata
This should be feasible because only non-directories could have been
affected by the fix in 0.31 that added empty Metadata entries for
paths whose metadata couldn't be read from the filesystem after they'd
already been added to the pending tree:

  47891d8951a95b8e0d9ca94387107cdf12ca3d3c
  save: add empty metadata if reading fails

which fixed:

  16f9f9829038f25aec80ebfae3c882a66281e145
  save-cmd.py: don't crash when a path disappears between index and save.

i.e. an empty metadata entry for a dir is normal (i.e. in the bupm for
the dir's parent), and for a non-dir always means the metadata was
lost.

We also don't need to agument the rewrite db's key because the
restricted mode is in the item.mode, which we'll still have when
querying the db. The rewrite db only needs to track the content, and
whether or not the replacement was chunked (the latter so we'll know
what kind of git object to create).

tree._dir_metadata still reverses the process, since it converts all
integer modes back to empty metadata.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>