Change split tree representation to encode info in the path
Rather than denote a split tree by a top-level .bupd.DEPTH.bupd marker
file, encode the information in a suffix to every split subtree name,
e.g.

  dir/j..5.bupd/jo..4.bupd/...
  ...
  dir/s..5.bupd/st..4.bupd/...
  ...

Compression should make the storage impact negligible, and this makes
it possible to understand the structure of a path from the path alone,
i.e. without needing access to the repo in order to search trees for
the .bupd.DEPTH.bupd files.

Signed-off-by: Johannes Berg <johannes@sipsolutions.net>
Reviewed-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>