get: improve symlink repairs and add tests
Handle more cases with respect to missing or incorrect symlink
information. Previously, we only handled cases where there was no
information at all.

When there's no information at all, continue replacing that with a
repair file and requiring --repair.

Fix the previous code to "restore" a symlink blob from the
metadata (when we have the latter but not the former). Previously it
was unreachable because vfs.readlink() doesn't throw
MissingObject. When we do "restore", treat it as a repair, reporting a
Bup-Repair-ID id, adding a Bup-Restored-Link trailer, and exiting with
status 3.

Since it's straightforward, detect the presumably rare situation where
the symlink target in the metadata doesn't match the target in the
blob. Refuse to fix it without --repair (i.e. repairs.destructive)
because we can't change that tree while we're including trees in the
rewrite db --- just mention --repair and EXIT_FAILURE for now. When
repairing, replace the path's blob with the metadata's target and add
a Bup-Fixed-Link-Blob trailer.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>