rewrite: don't remember any trees when there are excludes
When an exclude applies, it changes the tree, and since the rewrite db
is shared across all of a get's targets we must not store entries for
any such trees. Otherwise a rewrite --copy following a --rewrite might
find and use a tree inserted by the --rewrite that had exclusions.

Since an exclusion also affects all of the tree's parents, for now
just treat this case like --repair, and don't remember *any* trees.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>