test-import-duplicity: explicitly disable encryption
Invoke duplicity with --no-encryption to avoid failures since the
tests don't need or want encryption, and without this the test failed
on a FreeBSD host in CI with a passphrase related error. Could be that
somehow PASSPHRASE was set in the environment there.

Signed-off-by: Rob Browning <rlb@defaultvalue.org>
Tested-by: Rob Browning <rlb@defaultvalue.org>
(cherry picked from commit a056dd765d9b51a8245c4066b3937dbe37174582)